refactor: move helpers to a single file and extract commong logic
