+++ handlers/messages/sunday.go
@@ -1,4 +1,4 @@
-package handlers
+package messages
 
 import (
 	"fmt"
@@ -10,19 +10,13 @@ import (
 )
 
 var (
-	sundayTrigger  = "domingo a noite"
-	sundayGifUrl   = "https://media1.tenor.com/m/3udM407rgkQAAAAC/kikimogi-kiki.gif"
+	sundayTrigger        = "domingo a noite"
+	sundayGifUrl         = "https://media1.tenor.com/m/3udM407rgkQAAAAC/kikimogi-kiki.gif"
 	sundayFallbackGifUrl = "https://media1.tenor.com/m/UAK1t55SoRUAAAAd/chaves-chavo.gif"
 )
 
-
 func SundayHandler(s *discordgo.Session, m *discordgo.MessageCreate) {
 	cfg = config.GetConfig()
-
-	if m.Author.ID == s.State.User.ID {
-		return
-	}
-
 	if !strings.Contains(strings.ToLower(m.Content), sundayTrigger) {
 		return
 	}
@@ -36,7 +30,7 @@ func SundayHandler(s *discordgo.Session, m *discordgo.MessageCreate) {
 		content = "É hoje... Ele chegou..."
 		gif = getRandomGif("o malvado")
 		gifUrl = extractGifFromGTenor(gif, sundayGifUrl)
-		
+
 	case time.Saturday:
 		content = "É amanhã... Ele nunca esteve tão perto..."
 		gif = getRandomGif("prepare")
@@ -48,22 +42,12 @@ func SundayHandler(s *discordgo.Session, m *discordgo.MessageCreate) {
 		gifUrl = extractGifFromGTenor(gif, sundayFallbackGifUrl)
 
 	default:
-		content = fmt.Sprintf("Prepare-se, o evento canônico ocorrerá em %d dia(s)... Domingo a noite se aproxima.", daysRemainingToFriday())
+		content = fmt.Sprintf("Prepare-se, o evento canônico ocorrerá em %d dia(s)... Domingo a noite se aproxima.", daysReminingTo(time.Sunday))
 		gif = getRandomGif("rezem")
 		gifUrl = extractGifFromGTenor(gif, sundayFallbackGifUrl)
 	}
 
-	message := &discordgo.MessageSend{
-		Content: content,
-		Files:   []*discordgo.File{processGifUrl(gifUrl)},
-		Reference: &discordgo.MessageReference{
-				MessageID: m.ID,
-				ChannelID: m.ChannelID,
-				GuildID:   m.GuildID,
-		},
-	}
-
-	_, err := s.ChannelMessageSendComplex(m.ChannelID, message)
+	err := replyMessage(m, s, gifUrl, content)
 
 	if err != nil {
 		fmt.Println("Sunday Handler - There was an exception when sending a message", err)
